implementation and tests

+++ models/nodepoolcost_dto.go
@@ -4,7 +4,7 @@ import "time"
 
 type NodePoolCostDto struct {
 	Id         int32       `gorm:"column:id;type:int;primaryKey;<-:false"`
-	Cost       int32       `gorm:"column:cost;type:int;<-:false"`
+	Cost       float64     `gorm:"column:cost;type:int;<-:false"`
 	Currency   string      `gorm:"column:cost_currency;type:char(3);<-:false"`
 	FromDate   time.Time   `gorm:"column:from_date;type:datetimeoffset(0);<-:false"`
 	ToDate     time.Time   `gorm:"column:to_date;type:datetimeoffset(0);<-:false"`
